+++ src/Layout/Nav/Nav.jsx
@@ -3,8 +3,6 @@ import { Link } from 'react-router-dom';
 import schoolLogo from '../../assets/images/gnu-logo.png';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faChevronDown } from '@fortawesome/free-solid-svg-icons';
-import Modal from '../../components/Modal/Modal';
-import Login from '../../pages/Login/Login';
 
 const navItems = [
     { kr: '홈', en: 'Home', link: '/' },
@@ -15,10 +13,9 @@ const navItems = [
     { kr: '문의', en: 'Contact', link: '/contact' },
 ];
 
-const Nav = () => {
+const Nav = ({ openLoginModal }) => {
     const [toggle, setToggle] = useState(false);
     const [lang, setLang] = useState('KR');
-    const [isLoginModalOpen, setIsLoginModalOpen] = useState(false);
 
     const handleLangSelectToggle = () => {
         setToggle((prevToggle) => !prevToggle);
@@ -29,9 +26,6 @@ const Nav = () => {
         setLang(lang);
         setToggle(false);
     };
-
-    const openLoginModal = () => setIsLoginModalOpen(true);
-    const closeLoginModal = () => setIsLoginModalOpen(false);
 
     return (
         <div className="fixed top-0 left-0 w-full flex items-center justify-around py-5 bg-white shadow-md z-50">
@@ -77,9 +71,6 @@ const Nav = () => {
                     </div>
                 </div>
             </div>
-            <Modal isOpen={isLoginModalOpen} onClose={closeLoginModal}>
-                <Login onClose={closeLoginModal} />
-            </Modal>
         </div>
     );
 };
