--- src/pages/Signup/Signup.jsx
@@ -1,9 +1,0 @@
-const Signup = () => {
-    return (
-        <div className="flex pt-28 pb-10">
-            <div className="text-4xl font-bold py-8 border-b-2 border-black w-full m-8 text-center">회원 가입</div>
-        </div>
-    );
-}
-
-export default Signup;
