+++ src/routes/routes.jsx
@@ -1,22 +1,32 @@
 import '../assets/styles/tailwind.css';
+import { useState } from 'react';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import Nav from '../Layout/Nav/Nav';
 import Main from '../pages/Main/Main';
 import Publication from '../pages/Publication/Publication';
-import Login from '../pages/Login/Login';
-import Signup from '../pages/Signup/Signup'
+import Sign from '../pages/Sign/Sign';
+import Modal from '../components/Modal/Modal';
 import Footer from '../Layout/Footer/Footer';
 
 function App() {
+    const [isLoginModalOpen, setIsLoginModalOpen] = useState(false);
+
+    const openLoginModal = () => setIsLoginModalOpen(true);
+    const closeLoginModal = () => setIsLoginModalOpen(false);
+
     return (
         <BrowserRouter>
-            <Nav />
+            <Nav openLoginModal={openLoginModal} />
             <Routes>
                 <Route path="/" element={<Main />} />
                 <Route path="/publication" element={<Publication />} />
-                <Route path="/signin" element={<Login />} />
-                <Route path='/signup' element={<Signup/>}/>
+                <Route path="/sign" element={<Sign />} />
             </Routes>
+            {isLoginModalOpen && (
+                <Modal isOpen={isLoginModalOpen} onClose={closeLoginModal}>
+                    <Sign />
+                </Modal>
+            )}
             <Footer />
         </BrowserRouter>
     );
